+++ packages/frontend/src/i18n/en-US/index.ts
@@ -40,6 +40,8 @@ export default {
     passwordsDontMatch: 'Passwords do not match',
     emailOrUsername: 'Email or Username',
     rememberMe: 'Remember me',
+    banned: 'Your account has been banned',
+    deleted: 'This account no longer exists',
   },
   common: {
     loading: 'Loading...',